Split key/ into a folder; merge ActiveFileRefKey/ActiveRef; prune trivial tests
- registry/key.rs promoted to key/ folder module: composite.rs,
  run_doc.rs, versioned.rs, retention.rs. Reads and encodes one key
  shape per file.
- ActiveFileRefKey + ActiveRef merge into one ActiveFileRef in
  retention/active_refs.rs. The row IS the key (empty value slot);
  three-Uuid struct + to_bytes(self) for fjall + parse(bytes) for
  reads. Removes the old parse-returns-tuple / new-returns-key
  asymmetry and drops one type.
- Cut 17 low-value tests: trivial round-trips, idempotency-of-
  fjall-primitives, one-line invariants (Zero deadline == pinned_at,
  Duration deadline adds days, Indefinite yields no row).
  Kept every multi-tenant / scope-disambiguation / deadline-filter /
  strictness-ordering / reap-branch / defensive-parse test.
