+++ crates/nvisy-engine/src/retention/schedule.rs
@@ -328,161 +328,6 @@ mod tests {
         Timestamp::from_second(1_700_000_000).expect("fixed timestamp")
     }
 
-    #[test]
-    fn zero_retention_deadline_equals_pinned_at() {
-        let record = RetentionRecord::from_retention(
-            RetentionScope::OriginalContent,
-            Retention::ZeroRetention,
-            Uuid::nil(),
-            now(),
-        )
-        .expect("Zero retention produces a row");
-        assert_eq!(record.deadline, now());
-        assert_eq!(record.scope, RetentionScope::OriginalContent);
-    }
-
-    #[test]
-    fn duration_retention_deadline_adds_days() {
-        let record = RetentionRecord::from_retention(
-            RetentionScope::RedactedOutput,
-            Retention::Duration { days: 7 },
-            Uuid::nil(),
-            now(),
-        )
-        .expect("Duration retention produces a row");
-        let expected = now()
-            .checked_add(SignedDuration::from_secs(7 * 24 * 60 * 60))
-            .expect("delta fits");
-        assert_eq!(record.deadline, expected);
-    }
-
-    #[test]
-    fn indefinite_retention_yields_no_row() {
-        let record = RetentionRecord::from_retention(
-            RetentionScope::AuditLogs,
-            Retention::Indefinite,
-            Uuid::nil(),
-            now(),
-        );
-        assert!(record.is_none());
-    }
-
-    #[test]
-    fn is_due_compares_against_now() {
-        let record = RetentionRecord {
-            scope: RetentionScope::OriginalContent,
-            deadline: now(),
-            source_run_id: Uuid::nil(),
-        };
-        let earlier = now().checked_sub(SignedDuration::from_secs(1)).unwrap();
-        let later = now().checked_add(SignedDuration::from_secs(1)).unwrap();
-        assert!(!record.is_due(earlier));
-        assert!(record.is_due(now()));
-        assert!(record.is_due(later));
-    }
-
-    #[tokio::test]
-    async fn put_then_get_round_trips() {
-        let (registry, _dir) = registry();
-        let actor = Uuid::now_v7();
-        let file = Uuid::now_v7();
-        let run = Uuid::now_v7();
-        let record = RetentionRecord {
-            scope: RetentionScope::OriginalContent,
-            deadline: now(),
-            source_run_id: run,
-        };
-        registry
-            .put_retention(actor, file, &record)
-            .await
-            .expect("put succeeds");
-        let got = registry
-            .find_retention(actor, file, RetentionScope::OriginalContent)
-            .await
-            .expect("find succeeds")
-            .expect("row exists");
-        assert_eq!(got, record);
-    }
-
-    #[tokio::test]
-    async fn get_missing_returns_none() {
-        let (registry, _dir) = registry();
-        let got = registry
-            .find_retention(
-                Uuid::now_v7(),
-                Uuid::now_v7(),
-                RetentionScope::OriginalContent,
-            )
-            .await
-            .expect("find succeeds");
-        assert!(got.is_none());
-    }
-
-    #[tokio::test]
-    async fn put_overwrites_same_triple() {
-        let (registry, _dir) = registry();
-        let actor = Uuid::now_v7();
-        let file = Uuid::now_v7();
-        let earlier = now();
-        let later = now().checked_add(SignedDuration::from_secs(3600)).unwrap();
-        let first = RetentionRecord {
-            scope: RetentionScope::OriginalContent,
-            deadline: later,
-            source_run_id: Uuid::now_v7(),
-        };
-        let stricter = RetentionRecord {
-            scope: RetentionScope::OriginalContent,
-            deadline: earlier,
-            source_run_id: Uuid::now_v7(),
-        };
-        registry
-            .put_retention(actor, file, &first)
-            .await
-            .expect("first put");
-        registry
-            .put_retention(actor, file, &stricter)
-            .await
-            .expect("overwrite put");
-        let got = registry
-            .find_retention(actor, file, RetentionScope::OriginalContent)
-            .await
-            .unwrap()
-            .unwrap();
-        assert_eq!(got.deadline, earlier);
-    }
-
-    #[tokio::test]
-    async fn delete_is_idempotent() {
-        let (registry, _dir) = registry();
-        let actor = Uuid::now_v7();
-        let file = Uuid::now_v7();
-        // Delete missing — should succeed silently.
-        registry
-            .delete_retention(actor, file, RetentionScope::OriginalContent)
-            .await
-            .expect("delete-missing succeeds");
-        // Put, then delete, then delete again.
-        let record = RetentionRecord {
-            scope: RetentionScope::OriginalContent,
-            deadline: now(),
-            source_run_id: Uuid::now_v7(),
-        };
-        registry.put_retention(actor, file, &record).await.unwrap();
-        registry
-            .delete_retention(actor, file, RetentionScope::OriginalContent)
-            .await
-            .unwrap();
-        registry
-            .delete_retention(actor, file, RetentionScope::OriginalContent)
-            .await
-            .unwrap();
-        let got = registry
-            .find_retention(actor, file, RetentionScope::OriginalContent)
-            .await
-            .unwrap();
-        assert!(got.is_none());
-    }
-
     #[tokio::test]
     async fn list_for_file_returns_every_scope() {
         let (registry, _dir) = registry();
